requirements.txt
asgiref==3.4.0
bahttext==1.0.2
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-crispy-forms==1.12.0
django-storages==1.11.1
django-thaidate==1.0.2
gunicorn==20.1.0
idna==2.10
mysqlclient==1.4.6
Pillow==8.2.0
psycopg2==2.9.1
pythainlp==2.3.1
python-crfsuite==0.9.7
pytz==2021.1
reportlab==3.5.68
requests==2.25.1
selenium==3.141.0
sqlparse==0.4.1
thaidate==0.2.2
tinydb==4.5.0
urllib3==1.26.6
webdriver-manager==3.4.2
whitenoise==5.2.0
numpy==1.20.3
pandas==1.2.4